requirements.txt
djangorestframework==3.11.0
Django==3.0.2
sqlparse==0.3.0
drf_yasg
psycopg2